KUDU-2055 [part 1]: Add fs::BlockDeletionTransaction API
Similar to 'BlockCreationTransaction', this patch adds a new layer of
abstraction at Block Manager to coalesce blocks deletions in a logical
operation, e.g. compaction.

By coalescing blocks deletions, the number of holes punched in LBM will
be reduced from one per block to one per log block container in the
best case (that the individual holes in one container are all
contiguous). This should overall optimize the performance of operation
that involves batch deletions, such as tablet deletion and hole
repunching.

This patch is the first part of a series. It only adds the new
abstraction 'BlockDeletionTransaction', and doesn't yet change any
deletion semantics or improve performance.

Change-Id: Iecb252b3a5665d5471bb82301d0c8012a68de959
Reviewed-on: http://gerrit.cloudera.org:8080/7656
Tested-by: Kudu Jenkins
Reviewed-by: Dan Burkert <danburkert@apache.org>