KUDU-2055 [part 2]: Add util to construct sorted disjoint interval list
This patch adds a utility to construct a sorted disjoint interval list
in-place given a list of intervals. The operation to construct such one
is O(nlg n + n) where 'n' is the number of intervals. This util can be
used when log block manager coalesces block deletions belonging to the
same container.

For example, given the input interval list:
   [------2-------)         [-----1-----)
       [--3--)    [---5--)    [----4----)

The output sorted disjoint interval list is:
   [----------1----------)  [-----2-----)

It also adds unit test to verify given overlapped, duplicate, invalid
intervals, the implementation works as expected.

Change-Id: I61a813c047be4882f246eaf404598e7e18fcac87
Reviewed-on: http://gerrit.cloudera.org:8080/8041
Tested-by: Kudu Jenkins
Reviewed-by: Dan Burkert <danburkert@apache.org>