tablet: introduce closed mvcc and stopped tablets
Currently, the only way to stop an Applying transaction is to wait for
it to finish and Commit it. This constraint was put in place to
guarantee on-disk correctness, but is sometimes too strict. E.g. if the
tablet is shutting down, the Apply doesn't need to finish.

This patch adds a new state to the MvccManager in which it is closed
for transactions. Once in this closed state:
1. New Applies will return and not move to the Commit phase, and any
   methods waiting for the tablet's Applies to Commit (e.g. new snapshot
   scans, FlushMRS) will respond with an error immediately. This allows
   an escape from the existing invariant that Applies _must_ Commit,
   provided the MvccManager is in this closed state.
2. Applies that are already underway may still Commit, but will return
   early on a best-effort basis. These non-Committed operations are
   inconsequential w.r.t. consistency; having some in-flight
   transactions Commit and others not is consistent with the server
   shutting down in between the Commits of two transactions.
3. New transactions drivers will abort immediately before even reaching
   the Prepare phase, ensuring no more writes to the tablet are made
   durable.

The Tablet class uses this closed MVCC state in a new "stopped" state of
its own. A Tablet that has been stopped will avoid further activity:
its MvccManager is closed to prevent further writes, and its maintenance
ops are cancelled to prevent further scheduling.

This patch includes these new behaviors when shutting down a tablet,
with the assumption that a tablet will only be shut down when it's being
deleted and we don't care too much about its in-flight transactions
Committing or its further maintenance ops. Code paths that previously
crashed if Applies did not succeed (e.g. TransactionDriver::ApplyTask,
MvccManager::AbortTransaction, etc.) or that waited for Applies to
finish (e.g. Tablet:: FlushUnlocked) will now _not_ crash if the Tablet
has been stopped and will log a warning instead.

Testing is done by adding the following:
- a test in mvcc-test to shut down MVCC and delete an Applying
  transaction, ensuring that there are no errors when it leaves scope.
- a test in mvcc-test to wait on an Applying transaction, shut down
  MVCC, and ensure that any waiters will return with an error.
- a new test stop_tablet-itest is added to ensure stopped leaders block
  writes (because they cannot start new transactions) and stopped
  followers don't (because while they cannot service the op, there still
  exists a majority that can); that stopped tablets don't prevent
  fault-tolerant scans; and that stopping the only tablet does prevent
  scans

Change-Id: I983620f27e7226806a2cca253db7619731914d42
Reviewed-on: http://gerrit.cloudera.org:8080/7439
Reviewed-by: Mike Percy <mpercy@apache.org>
Tested-by: Kudu Jenkins
Reviewed-by: Todd Lipcon <todd@apache.org>