[catalog manager] introduce replica type filter
Introduced a replica type filter for GetTabletLocations RPC and
various wrapper methods used in master, catalog manager, and tests.
The filter is used to specify what type of tablet replicas to include
while populating the result of the GetTabletLocations RPC.

When not specified otherwise or when using earlier versions of the
GetTable[t]LocationsRequestPB protobuf messages, VOTER_REPLICA policy
is used.  To locate all replicas, specify ANY_REPLICA policy in the
'replica_type_filter' field of the GetTable[t]LocationsRequestPB.  The
latter is done now unconditionally for already existing non-voter
replica-related tests.  This will change by a follow-up commit:
the 'regular' client operations should work with voter replicas only,
so the use cases for getting all tablet replicas will be just
the following:

  * the kudu CLI tool fetching that information to report on replica
    membership status for listings in 'table list --tablets' and
    'ksck' sub-commands.

  * various tests exercising the functionality related to
    non-voter tablet replicas

Added an integration test to verify the functionality of the replica
type filter for GetTableLocations and GetTabletLocations RPC.

Change-Id: I303a6d158184575a9a105c2d2bf26961ae8b3e93
Reviewed-on: http://gerrit.cloudera.org:8080/8161
Tested-by: Alexey Serbin <aserbin@cloudera.com>
Reviewed-by: Mike Percy <mpercy@apache.org>
Reviewed-on: http://gerrit.cloudera.org:8080/8727
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>
Tested-by: Mike Percy <mpercy@apache.org>