mini-cluster: new module for the mini cluster implementations
For the upcoming "control shell" patch, the Kudu CLI will need to link
ExternalMiniCluster and friends. Today that means depending on the
integration-tests target, which doesn't feel quite right since that's
mostly, well, integration tests. More practically, the integration-tests
cmake target is conditioned on !NO_TESTS but the Kudu CLI is not.

So this patch moves the mini cluster code into a standalone module. I also
took the opportunity to place the code in a new "cluster" namespace.

Change-Id: I0437e281da5874016d9c1f1404a6de043bfb4088
Reviewed-on: http://gerrit.cloudera.org:8080/8173
Tested-by: Kudu Jenkins
Reviewed-by: Dan Burkert <danburkert@apache.org>