external_mini_cluster: don't pipe daemon subprocess stdout
I don't know why ExternalDaemon was configured like this, since nobody
appears to be using the subprocess' stdout pipe. It's actually pretty
dangerous: if the subprocess were to write over 64K of data to stdout, it'd
fill the pipe and the next write to stdout would block indefinitely since no
one is reading from it.

I snuck in some cosmetic changes to {ExternalMiniCluster,MiniKdc}Options.

Change-Id: Iddd8d78cd085975821207d324ea2fd5365a2c2ba
Reviewed-on: http://gerrit.cloudera.org:8080/8187
Tested-by: Kudu Jenkins
Reviewed-by: Todd Lipcon <todd@apache.org>