mini-cluster: condition on NO_TESTS=0
This module will be incorporated into the kudu CLI in the near future, but
for now this patch fixes builds that set NO_TESTS=1.

Our precommit builds don't disable tests, so I tested this by running a
local build with NO_TESTS=1. Without this patch, the build failed while
linking libmini-cluster.so. With it, the build passes.

Change-Id: I195b3ed5ed50dfbe7afb3004182bf02a3bcda31a
Reviewed-on: http://gerrit.cloudera.org:8080/8199
Tested-by: Adar Dembo <adar@cloudera.com>
Tested-by: Kudu Jenkins
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>