KUDU-2167: fix C++ client crash due to bad assumption regarding scan data
The new unit test triggered the crash reliably, though it's probably not the
only way (or perhaps not even the best way) to trigger it.

I also modified the tserver to always populate the scan data fields so that
older clients are also protected. This necessitated an additional change to
how a scanner's "data in open" was computed: the existence of the 'data'
field in the scan response isn't enough. The Java client was already doing
this, so it seems like a reasonable change to make on the C++ side.

Change-Id: If1a8a4e22082cf39710b9f00894f644a0b34234e
Reviewed-on: http://gerrit.cloudera.org:8080/8204
Reviewed-by: Todd Lipcon <todd@apache.org>
Tested-by: Kudu Jenkins
(cherry picked from commit 858bf73b380d10f16e7aca04a780a0adae705e9a)
Reviewed-on: http://gerrit.cloudera.org:8080/8224
Tested-by: Adar Dembo <adar@cloudera.com>
Reviewed-by: Jean-Daniel Cryans <jdcryans@apache.org>