KUDU-2161 ColumnSchemaBuilder missing Audience annotation
ColumnSchema.ColumnSchemaBuilder is missing the Yetus Audience
and Stability annotations. This gives it an implied private
audience annotation, so Yetus's IncludePublicAnnotationsStandardDoclet
is omitting its docs from javadoc.

This applies ColumnSchema's annotations (Public + Evolving) to
ColumnSchemaBuilder, which is nested inside ColumnSchema.

I verified that the generated javadocs include ColumnSchemaBuilder
with this change.

Change-Id: I51576db7e28033eec9378dac238d28725aca1921
Reviewed-on: http://gerrit.cloudera.org:8080/8214
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Kudu Jenkins