Add log parser script
This script collates and filters logs from a cluster, printing messages
indicating latency issues and RaftConsensus election activity. It's a
work-in-progress but I found it useful for an investigation I was doing
(it highlighted some problems I hadn't noticed).

It's pretty simple and basically consists of a bunch of regexes it runs
on the logs. It's also a little slow and could be optimized, as it
currently does an in-memory sort (by timestamp) of any included log
entries right before it prints the relevant log lines to stdout.

Change-Id: I3bd8bb5d9a3c598ade7bc1bbc8f5a9e24ca618af
Reviewed-on: http://gerrit.cloudera.org:8080/8229
Tested-by: Kudu Jenkins
Reviewed-by: Will Berkeley <wdberkeley@gmail.com>