KUDU-2187. Don't hold threadpool lock while creating threads
This changes the threadpool implementation so that if a task submission
needs to start a thread, it does so after releasing the lock. This
enables other worker threads to continue to process items even if thread
creation is slow.

This is meant to address various cases where I've seen stalled writes
with high 'mutex_wait_us' metrics. Looking at the logs it is sometimes
the case that this correlates with a single slow pthread creation.

Change-Id: If91cb032db25ed539ec8a952f302cf4501b3c240
Reviewed-on: http://gerrit.cloudera.org:8080/8256
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Kudu Jenkins