tsan: flush TSAN memory every 10 seconds
I'm seeing a TSAN precommit test fail in a hard-to-diagnose way because the
Linux OOM killer is killing tservers underneath it. It doesn't appear to
be any leak -- rather, TSAN is just using too much memory due to the
test doing a lot of allocation.

This enables a TSAN option to flush its memory every 10 seconds. This
has the downside of potentially missing races that happen far apart in
time (but those are unlikely to actually trigger in practice) but
reduces RSS significantly.

Change-Id: I721f126eb731c7f35afb732cebe48abd52c3ebd1
Reviewed-on: http://gerrit.cloudera.org:8080/8251
Tested-by: Alexey Serbin <aserbin@cloudera.com>
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>