catalog manager: fix DDL race
This fixes a race between concurrent ALTER TABLE RENAME and DROP TABLE
DDL operations. Previously it was possible for two such operations to
concurrently complete on the same table. The included test is about 50%
flaky without the fix.

In addition, it changes the table finding logic to use the table ID and
table name if both are present in the TableIdentifier. Previously only
the table ID would be used in this case. No clients currently send both,
so this shouldn't have any real effect, but I think it's useful to be
able to specify both in certain cases.

Change-Id: I84ca3b207da28cd7dc43a077736da9b4e0ec6f37
Reviewed-on: http://gerrit.cloudera.org:8080/8254
Tested-by: Kudu Jenkins
Reviewed-by: Dan Burkert <danburkert@apache.org>