arena: remove second parameter for max buffer size
It was never very obvious how to pick a value for the max buffer size of
an Arena, and upon some reflection I think 1MB is almost always the
right choice. So, this removes the parameter for it from the
constructor, and instead adds a setter for the rare case when it needs
to be customized.

Change-Id: I56d7bb5b689c608382423038f11013b6be39e656
Reviewed-on: http://gerrit.cloudera.org:8080/8269
Tested-by: Kudu Jenkins
Reviewed-by: Todd Lipcon <todd@apache.org>