[client-stress-test] do not crash if client failed
Do not crash the test if failed to build Kudu client instance in the
TestUniqueClientIds scenario, just use regular ASSERT_OK() instead,
allowing the test cluster to wind down, so it would be possible to
check on the status of the master and tablet servers during shutdown.

With crashing the test on an error, it's not possible to figure out
whether client <--> server negotiation failed due to a crashed server
or something else.

As for the context, the ClientStressTest.TestUniqueClientIds test
has been flaky for a while, failing with errors like the following:

  Client connection negotiation failed: client connection to IP:port: \
    Read zero bytes on a blocking Recv() call: Transferred 0 of 4 bytes

Change-Id: I558fcb02f697d1ec216538416826d96dcda32a1e
Reviewed-on: http://gerrit.cloudera.org:8080/8310
Tested-by: Alexey Serbin <aserbin@cloudera.com>
Reviewed-by: Adar Dembo <adar@cloudera.com>