kudu-hive: add provided scope dependency on hadoop-mapreduce-client-common
This appears to be necessary to get the kudu-hive tests to run against
some versions of Hadoop, otherwise NoClassDefFound exceptions occur for
the following classes while attempting to create the Hive configuration
object:

- org.apache.hadoop.mapreduce.TaskAttemptContext
- org.apache.hadoop.mapred.MRVersion

I haven't untangled the chain of Hadoop modules and Maven dependencies
to figure out exactly why this occurs, but since the fix is harmless
enough I'm inclined to keep my sanity and let it lie.

Note: this isn't a problem with Gradle.

Change-Id: I62d62ae31ebd82a2fd39030bfa6201aef6ff82ff
Reviewed-on: http://gerrit.cloudera.org:8080/8316
Tested-by: Kudu Jenkins
Reviewed-by: Jean-Daniel Cryans <jdcryans@apache.org>