KUDU-1957: Clarify web UI redaction in --redact flag help
This patch refines the definition of --redact flag to control the
context of redaction. The available options now are 'all', 'log'
and 'none'. If 'all' is specified, sensitive data (sensitive
configuration flags and row data) will be redacted from the
web UI as well as glog and error messages. If 'log' is specified,
sensitive data will only be redacted from glog and error messages.
And the option 'flag' is removed from --redact flag to let this
flag only control redaction context.

Change-Id: I32fe0c4e65f22d6f9f3d188018b119f07300e26f
Reviewed-on: http://gerrit.cloudera.org:8080/6755
Tested-by: Kudu Jenkins
Reviewed-by: Dan Burkert <danburkert@apache.org>