tool: remove dependency on libcurl
This dependency was introduced in commit 761ce10, when I added support for
spawning mini clusters through the CLI tool. We could treat it as a valid
new dependency and doc it, but it only exists because of one ExternalDaemon
test function. So instead, let's move that function from ExternalDaemon to
the group of free functions in cluster_itest_util.

I verified that with this patch the CLI no longer depends on kudu_curl_util
or on libcurl, and that the build still passes with NO_TESTS=1.

Change-Id: Ifcf0b4eeed9f8af38cc471d81eb59a0473f470c3
Reviewed-on: http://gerrit.cloudera.org:8080/8364
Tested-by: Kudu Jenkins
Reviewed-by: Todd Lipcon <todd@apache.org>