fs: defer failure from metadata load to bootstrap when data dir is missing
An upcoming patch adds a CLI tool action to remove data directories. When a
data dir is removed, all tablets with data on it will fail. Today that
failure manifests as a FindOrDie in DataDirGroup::FromPB; we need to make
that a little bit more graceful.

This patch modifies the DataDirGroup FromPB/CopyToPB methods to return a
failure when a data dir is missing. It further changes TabletMetadata to
treat such failures as non-fatal, and adds checks to TabletBootstrap so that
the failures manifest there instead.

No tests in this patch because:
1. Andrew has already merged tablet-level tests for failed disks, and
2. The CLI tool patch adds coverage at the itest-level.

Change-Id: I1e8d5697c2bb08287cce11fbdab6fb8d6e37d1ad
Reviewed-on: http://gerrit.cloudera.org:8080/8376
Reviewed-by: Todd Lipcon <todd@apache.org>
Reviewed-by: Andrew Wong <awong@cloudera.com>
Tested-by: Kudu Jenkins