KUDU-1961: enable ccache when using devtoolset-3
The original approach outlined in the bug report doesn't work because
although /opt/rh/devtoolset-3/ is ahead of /usr/bin during
enable_devtoolset.sh, it isn't when make is invoked, so ccache ends up
using the old system compiler and the build fails.

So here's a more complicated approach: set CC/CXX to explicit ccache
"helper" scripts, in the style of those found in build-support/ccache-clang.
We do nothing if CC or CXX are already set; presumably the user knows better
than us.

In this manner, ccache is used by default in almost all cases. To opt out,
set CCACHE_DISABLE in the environment when running make.

Change-Id: Ic3a264fc9471e6486a3f17475697f1fa7b64a943
Reviewed-on: http://gerrit.cloudera.org:8080/8392
Tested-by: Adar Dembo <adar@cloudera.com>
Reviewed-by: Todd Lipcon <todd@apache.org>