thirdparty: bump crcutil hash
The new revision includes a patch that detects gcc correctly in more
circumstances. In my experience the detection (and build) would fail when
thirdparty was rebuilt by make directly, or when ccache was in use. This
should address that.

I also published the internal repo externally, something we should have done
a long time ago.

Change-Id: I60c2ad24ee6919ec1a1c7948c0ce5df468adde6b
Reviewed-on: http://gerrit.cloudera.org:8080/8390
Reviewed-by: Todd Lipcon <todd@apache.org>
Tested-by: Kudu Jenkins