KUDU-2200: provide better diagnostics when connecting to a subset of masters
This changes the ConnectToMaster RPC to send back the list of the master
peers, and then changes the clients to use this information to provide
better error messages in the case that the user has specified only a
subset of the live masters.

Note that this patch follows a "first do no harm" philosophy. In the
case that the user "gets lucky" and only specifies one of their three
masters, and that master happens to be the leader, we'll continue to
treat that as "successful".

Change-Id: I52f903e1aa5ae6948ca1ba6d4d856c3c9dc73d56
Reviewed-on: http://gerrit.cloudera.org:8080/8393
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>
Tested-by: Kudu Jenkins