[tools] limit per-session error buffer space
Introduced --error-buffer-size-bytes flag to control the size of the
per-session error buffer.  By default, the limit is set to 16MB per
session (i.e. per worker thread).

Prior to this patch, the per-session error buffer had no size limit
and could grow very large because of huge number of errors during the
run (e.g., because of duplicate value errors).

I also updated the in-line documentation to add an example of using
the tool for subsequent runs against the same table using the
--seq-start flag.

Change-Id: I2b7663c52e57b2d3a1e11b22301b804da784fa74
Reviewed-on: http://gerrit.cloudera.org:8080/8396
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Kudu Jenkins